Server/flask-kafka/agent1/producer/requirements.txt
flask
kafka-python

